requirements.txt
# Core pipeline
assemblyai
librosa
ffmpeg-python
python-dotenv
yt-dlp

# API (optional)
fastapi
uvicorn

# LLM
openai

# Testing
pytest
pytest-cov

# Static analysis
ruff
mypy
bandit
pre-commit
